Crate test_docs

The point of this crate is to be able to have enough different “kinds” of documentation generated so we can test each different features. Enable the feature some-feature to enjoy this crate even more! Enable the feature some-feature to enjoy this crate even more! Enable the feature some-feature to enjoy this crate even more!

Also, stop using bar as it’s deprecated. Also, stop using bar as it’s deprecated. Also, stop using bar as it’s deprecated.

Finally, you can use quz only on Unix or x86-64 . Finally, you can use quz only on Unix or x86-64 .

Re-exports

pub use ::std as TheStdReexport;

Modules

SidebarSort

cfgs

codeblock_sub

details

doc_block_table

fields

foreign_impl_order

huge_amount_of_consts

impls_indent

long_code_block

Very long code text hereIgoWithLongTextBecauseWhyNotAndWhyWouldntI.

long_code_block_link

Very long code text hereIgoWithLongTextBecauseWhyNotAndWhyWouldntI.

long_list

bla

notable

private

search_results

short_docs

tooltips

trait_bounds

trait_members

tyalias

Macros

a

attr_macro

An attribute bang macro.

b

decl_macro

derive_macro

A derive bang macro.

foo

heavily_documented_macro

Top-doc prose title

one_for_all_macro

repro

repro2

Structs

Foo

Just a normal struct.

HasEmptyTraits

HeavilyDocumentedStruct

Top-doc Prose title

ImplDoc

NotableStructWithLongName

StructWithPublicUndocumentedFields

TypeWithImplDoc

TypeWithNoDocblocks

UnderlyingFooBarBaz

Another type alias, this time with methods.

WarningStruct

I have warnings!

WithGenerics

ZyxwvutMethodDisambiguation

Enums

AnEnum

Some doc with code!

HeavilyDocumentedEnum

Top-doc Prose title

WhoLetTheDogOut

Just a normal enum.

Constants

CONST

Traits

AnotherOne

Woohoo! A trait!

EmptyTrait1

EmptyTrait2

EmptyTrait3

ItemsTrait

TraitWithNoDocblocks

TraitWithoutGenerics

ZyxwvutTrait

Functions

check_list_code_block

foo

Basic function with some code examples:

replaced_functionDeprecatedUnix

a thing with a label

safe_fn

some_more_function

Who doesn’t love to wrap a format! call?

unsafe_fn⚠

Type Aliases

SomeOtherTypeWithMethodsAndInlining

Another type alias, this time with methods.

SomeOtherTypeWithMethodsAndInliningAndTraits

SomeType

Just some type alias.

Unions

HeavilyDocumentedUnion

Top-doc prose title

Keywords

for

Some keyword.

Attribute Macros

attr_macro

An attribute bang macro.

one_for_all_macro

Derive Macros

derive_macro

A derive bang macro.

one_for_all_macro

Attributes

forbid

Some attribute.