test_docs

Module long_list

bla

Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque et libero ut leo interdum laoreet vitae a mi. Aliquam erat volutpat. Suspendisse volutpat non quam non commodo.

Praesent enim neque, imperdiet sed nisl at, lobortis egestas augue. Sed vitae tristique augue. Phasellus vel pretium lectus.

Praesent enim neque, imperdiet sed nisl at, lobortis egestas augue. Sed vitae tristique augue. Phasellus vel pretium lectus.

Praesent enim neque, imperdiet sed nisl at, lobortis egestas augue. Sed vitae tristique augue. Phasellus vel pretium lectus.

Another list:

TryFromBytes indicates that a type may safely be converted from certain byte sequence (conditional on runtime checks)

FromZeros indicates that a sequence of zero bytes represents a valid instance of a type

FromBytes indicates that a type may safely be converted from an arbitrary byte sequence