Crate test_docs

The point of this crate is to be able to have enough different “kinds” of documentation generated so we can test each different features. Enable the feature some-feature to enjoy this crate even more! Enable the feature some-feature to enjoy this crate even more! Enable the feature some-feature to enjoy this crate even more!

Also, stop using bar as it’s deprecated. Also, stop using bar as it’s deprecated. Also, stop using bar as it’s deprecated.

Finally, you can use quz only on Unix or x86-64 . Finally, you can use quz only on Unix or x86-64 .

Re-exports

pub use ::std as TheStdReexport;

Modules

codeblock_sub

details

doc_block_table

huge_amount_of_consts

keyword

Some keyword.

long_code_block

Very long code text hereIgoWithLongTextBecauseWhyNotAndWhyWouldntI.

long_code_block_link

Very long code text hereIgoWithLongTextBecauseWhyNotAndWhyWouldntI.

search_results

trait_members

Macros

a

b

foo

heavily_documented_macro

Top-doc prose title

repro

repro2

Structs

Foo

Just a normal struct.

HasEmptyTraits

HeavilyDocumentedStruct

Top-doc Prose title

NotableStructWithLongName

StructWithPublicUndocumentedFields

TypeWithImplDoc

TypeWithNoDocblocks

WithGenerics

Enums

AnEnum

Some doc with code!

HeavilyDocumentedEnum

Top-doc Prose title

WhoLetTheDogOut

Just a normal enum.

Constants

CONST

Traits

AnotherOne

Woohoo! A trait!

EmptyTrait1

EmptyTrait2

EmptyTrait3

TraitWithNoDocblocks

TraitWithoutGenerics

Functions

check_list_code_block

foo

Basic function with some code examples:

replaced_functionDeprecatedUnix

a thing with a label

safe_fn

some_more_function

Who doesn’t love to wrap a format! call?

unsafe_fn⚠

Type Definitions

SomeType

Just some type alias.

Unions

HeavilyDocumentedUnion

Top-doc prose title

Keywords

CookieMonster

Some keyword.